+++ resource/csdk/stack/test/stacktests.cpp
@@ -2529,11 +2529,11 @@ TEST_F(OCDiscoverTests, DISABLED_DiscoverResourceWithInvalidQueries)
 
 TEST(StackZoneId, getZoneId)
 {
-    uint32_t tempSize = 0;
+    size_t tempSize = 0;
     CAEndpoint_t *tempInfo = NULL;
     CAGetNetworkInformation(&tempInfo, &tempSize);
 
-    for (uint32_t i = 0; i < tempSize; i++)
+    for (size_t i = 0; i < tempSize; i++)
     {
         char *zoneId = NULL;
         EXPECT_EQ(OC_STACK_OK, OCGetLinkLocalZoneId(tempInfo[i].ifindex, &zoneId));
